pyproject.toml
[project]
name = "vwap-trader"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "docker>=7.1.0",
    "jupyterlab>=4.4.0",
    "nautilus-trader[docker,ib]>=1.216.0",
    "polars>=1.27.1",
    "python-dotenv>=1.1.0",
]
